refactor: rename old modules
